backend/requirements.txt
firebase-admin==6.6.0
openai==1.59.3
python-dotenv==1.0.1
feedparser==6.0.11
requests==2.32.3